+++ ksa_compliance/public/js/sales_invoice.js
@@ -21,7 +21,9 @@ frappe.require("/assets/ksa_compliance/js/update_invoice_mode_of_payment.js").th
             if (frm.doc.is_return){
                 is_advance_invoice(frm).then(is_advance => {
                     frm.set_df_property("update_outstanding_for_self", "read_only", is_advance ? 1 : 0);
-                    frm.set_value("update_outstanding_for_self", 1);
+                    if (is_advance === 1){
+                       frm.set_value("update_outstanding_for_self", 1);
+                    }
                     frm.refresh_field("update_outstanding_for_self");
                 });
             }